feat: AdapterFactory实现类命名调整
